:rainbow: 尝试支持 Generations 的规则，为此改变了输出格式；可能还有 bug

+++ lib/src/states.rs
@@ -12,7 +12,7 @@ pub const ALIVE: State = State(1);
 /// where `None` means that the state of the cell is unknown.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
-pub struct State(pub(crate) u8);
+pub struct State(pub(crate) usize);
 
 /// Flips the state.
 impl Not for State {
